refactor: merge container into AgentStatus; remove AgentStatusContainer
