+++ console/src/AgentStatus.test.tsx
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "bun:test";
 import { renderToStaticMarkup } from "react-dom/server";
-import { AgentStatus } from "./AgentStatusContainer";
+import { AgentStatus } from "./AgentStatus";
 import { createAgentStateWebSocketUrl } from "./agentStateService";
 import { GameStatusSection } from "./agentStatusSections/GameStatusSection";
 import { LiveDeliveryStatusSection } from "./agentStatusSections/LiveDeliveryStatusSection";
